apex: Add unix socket control interface
Add a unix socket control interface on linux. This interface uses
jsonrpc and supports checking the status apexd. It will report back
the info needed to register the device if it's waiting for auth. This
would be one way to get info when running apex as a daemon. The other
would be from the daemon logs.

The patch also adds apex, a binary to be installed alongside of apexd.
It can issue commands over a local socket to apexd. Right now the only
command is status.

We now embed a version at build time into each binary. This is done
using ldflags to set the value of main.Version.

apex and apexctl can report their own version. apex will also report
the version of a running apexd. When running subcommands other than
"version", apex will first validate that the apex and apexd versions
match.

Signed-off-by: Russell Bryant <rbryant@redhat.com>