proxy: Add --run-test-service
Add a built-in trivial tcp listener to give you something you can hit
with curl or a browser. It's not an http server at all, but responds
pretending to be one if the received data starts with GET /.
It's enough to hit it with curl or some other http client for a quick
test. You can also use nc or telnet against the service and it will
act as an echo server.

Signed-off-by: Russell Bryant <russell.bryant@gmail.com>